Some basic bot functionality
Moves back and forth across the screen. When it stops, it spins 360 degrees to scan for other bots.
If it sees another bot, it checks its distance. If the bot is too far away, it doesn't fire so as to not waste energy. If the bot is close, it determines how close, and based off of that value, it will adjust firing power.
If it is hit, it will turn so that it hopefully will not be hit so easily again.
If it runs into a wall, it will back away from the wall and turn.