Changed functionality drastically - much simpler
I left the old code in, but it's commented out in case I use it later.

Now, the bot moves forward 500 units, then turns 90, then moves forward again, rinse and repeat. Whenever it stops moving, it spins the scanner and gun 360 degrees. If it scans a bot, and that bot is too far away, it won't fire. If the bot is close enough, then it will fire. It will adjust the power of the shot depending on how close it is.